Simplify tokenization (and this happens to fix a bug)
The bug occurred with anonymized data if the voter had empty choices in positions 5 and 4. In this case, the two empty choices was merged, which offsetted the rest of the choices for this voter: (5&4)->5, 3->4, 2->3, 1->2, GOTY description->1. As the GOTY description was interpreted as a game name, no result could be found on IGDB, and the software would crash, waiting for a manual fix of the matches.