Data: anonymized 2025 CSV (without any review)
I had to replace a semi-colon ";" with a dot "." in the review of ballot n°11
so that a part of the review does not persist in the anonymized
output.

However, please note that this was not causing any issue with the parsing of the ballots, and the final results were the same without the manual replacement!